+++ assets/block-editor/data/messages-store/actions.jsx
@@ -1,0 +1,13 @@
+export function fetchMessages() {
+	return {
+		type: 'FETCH_MESSAGES',
+	};
+}
+
+export function receiveMessages( messages, error = null ) {
+	return {
+		type: 'RECEIVE_MESSAGES',
+		messages,
+		error,
+	};
+}
