+++ assets/block-editor/data/messages-store/name.jsx
@@ -1,0 +1,1 @@
+export const MESSAGES_STORE = 'sensei-lms/messages';
